backend/requirements.txt

fastapi
uvicorn
python-jose
python-multipart
python-dotenv
psycopg2-binary
pythermalcomfort
paho-mqtt